Back-end/requirements.txt
asgiref==3.6.0
azure-cognitiveservices-vision-customvision==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
whitenoise==6.4.0
